changed location id to be 32-bit instead of 64-bit. Makes encdoing wr_id have more information to store channel count

+++ tests/test_exchange_client/channel.h
@@ -33,7 +33,7 @@ typedef enum message_type {
 
 
 typedef struct bid_order {
-	uint64_t location_id;
+	uint32_t location_id;
 	uint64_t data_bytes;
 	// for BIDs upon a match
 	uint64_t wr_id;
@@ -41,13 +41,13 @@ typedef struct bid_order {
 } Bid_Order;
 
 typedef struct offer_order {
-	uint64_t location_id;
+	uint32_t location_id;
 	uint64_t data_bytes;
 	uint8_t fingerprint[FINGERPRINT_NUM_BYTES];
 } Offer_Order;
 
 typedef struct future_order {
-	uint64_t location_id;
+	uint32_t location_id;
 	uint64_t data_bytes;
 	uint8_t fingerprint[FINGERPRINT_NUM_BYTES];
 } Future_Order;
@@ -56,7 +56,7 @@ typedef struct future_order {
 // and we can locate the fingerprint based on outstanding bids table within client
 // where location_id is chosen offer order to relay
 typedef struct Bid_Match{
-	uint64_t location_id;
+	uint32_t location_id;
 } Bid_Match;
 
 
@@ -66,12 +66,12 @@ typedef struct Channel_Item {
 
 
 typedef struct channel {
-	uint64_t self_id;
-	uint64_t peer_id;
-	uint64_t message_size;
+	uint32_t self_id;
+	uint32_t peer_id;
+	uint32_t message_size;
 	MessageType message_type;
-	uint16_t capacity;
-	uint16_t cnt;
+	uint32_t capacity;
+	uint32_t cnt;
 	// will be of size capacity * message_size
 	void * buffer;
 	// will store wr_id's and will call "find_item_index_table(wr_id)"
@@ -98,8 +98,8 @@ typedef struct channel {
 // Because using UD and shared QPs from various senders (shared wr_ids) across various protocols, 
 // need to ensure that senders only consume recvs meant for them, and the receivers know how to post sends
 
-// 0 - 40: sender id (so as to ensure senders don't step on each other toes consuming the same recv wr_ids)
-// 40 - 56: channel count
+// 0 - 32: sender id (so as to ensure senders don't step on each other toes consuming the same recv wr_ids)
+// 32 - 56: channel count
 //				- Simply for internal use on receiver end to maintain recv buffer mappings between wr_id and ring_buffer memory addr.
 //				- Sender maintains their own channel count and increments by 1 for each message
 //					- the receiver's corresponding recv request will be already be waiting due to protocol
@@ -109,11 +109,11 @@ typedef struct channel {
 
 char * message_type_to_str(MessageType message_type);
 
-uint64_t encode_wr_id(uint64_t sender_id, uint64_t channel_count, MessageType message_type);
+uint64_t encode_wr_id(uint32_t sender_id, uint32_t channel_count, MessageType message_type);
 
-MessageType decode_wr_id(uint64_t wr_id, uint64_t * ret_sender_id);
+MessageType decode_wr_id(uint64_t wr_id, uint32_t * ret_sender_id);
 
-Channel * init_channel(uint64_t self_id, uint64_t peer_id, uint16_t capacity, MessageType message_type, uint64_t message_size, bool is_inbound, bool to_presubmit_recv, struct ibv_pd * pd, struct ibv_qp * qp, struct ibv_cq_ex * cq);
+Channel * init_channel(uint32_t self_id, uint32_t peer_id, uint32_t capacity, MessageType message_type, uint32_t message_size, bool is_inbound, bool to_presubmit_recv, struct ibv_pd * pd, struct ibv_qp * qp, struct ibv_cq_ex * cq);
 
 // For in channels
 int submit_in_channel_reservation(Channel * channel, uint64_t * ret_wr_id, uint64_t * ret_addr);
